First Commit
	Finished
If a game is won, canvas should stay enabled forever: FIXED
Created property for MainWinner backingArray
Allowed CanvasManager to access backingArray
If something on the backingArray is not 0, then keep it active

	In-Progress
Create a graph data structure to store the subgrids?
Some sort of way to track adjacency sub-grids to determine which subgrid the player would go next after a win.